FINANCE-AGENT/requirements.txt
langchain_community==0.3.21
langchain_groq==0.3.2
langgraph==0.3.29
langgraph_supervisor==0.0.16
phidata==2.7.10
python-dotenv==1.1.0
streamlit==1.44.1
duckduckgo_search==8.0.0
yfinance==0.2.55
langsmith
portkey_ai
